fix test import cycle
