+++ pkg/util/signaler_test.go
@@ -6,8 +6,6 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-
-	"github.com/kedacore/http-add-on/pkg/routing"
 )
 
 var _ = Describe("Signaler", func() {
@@ -27,7 +25,7 @@ var _ = Describe("Signaler", func() {
 		It("produces on channel", func() {
 			s := make(signaler, 1)
 
-			err := routing.withTimeout(time.Second, routing.deapplyError(s.Signal, nil))
+			err := WithTimeout(time.Second, DeapplyError(s.Signal, nil))
 			Expect(err).NotTo(HaveOccurred())
 
 			select {
@@ -41,7 +39,7 @@ var _ = Describe("Signaler", func() {
 			s := make(signaler, 1)
 			s <- struct{}{}
 
-			err := routing.withTimeout(time.Second, routing.deapplyError(s.Signal, nil))
+			err := WithTimeout(time.Second, DeapplyError(s.Signal, nil))
 			Expect(err).NotTo(HaveOccurred())
 		})
 	})
@@ -53,7 +51,7 @@ var _ = Describe("Signaler", func() {
 			s := make(signaler, 1)
 			s <- struct{}{}
 
-			err := routing.withTimeout(time.Second, routing.applyContext(s.Wait, ctx))
+			err := WithTimeout(time.Second, ApplyContext(s.Wait, ctx))
 			Expect(err).NotTo(HaveOccurred())
 		})
 
@@ -63,7 +61,7 @@ var _ = Describe("Signaler", func() {
 
 			s := make(signaler, 1)
 
-			err := routing.withTimeout(time.Second, routing.applyContext(s.Wait, ctx))
+			err := WithTimeout(time.Second, ApplyContext(s.Wait, ctx))
 			Expect(err).To(MatchError(context.Canceled))
 		})
 	})
@@ -75,18 +73,18 @@ var _ = Describe("Signaler", func() {
 
 			s := NewSignaler()
 
-			err0 := routing.withTimeout(time.Second, routing.deapplyError(s.Signal, nil))
+			err0 := WithTimeout(time.Second, DeapplyError(s.Signal, nil))
 			Expect(err0).NotTo(HaveOccurred())
 
-			err1 := routing.withTimeout(time.Second, routing.deapplyError(s.Signal, nil))
+			err1 := WithTimeout(time.Second, DeapplyError(s.Signal, nil))
 			Expect(err1).NotTo(HaveOccurred())
 
-			err2 := routing.withTimeout(time.Second, routing.applyContext(s.Wait, ctx))
+			err2 := WithTimeout(time.Second, ApplyContext(s.Wait, ctx))
 			Expect(err2).NotTo(HaveOccurred())
 
 			cancel()
 
-			err3 := routing.withTimeout(time.Second, routing.applyContext(s.Wait, ctx))
+			err3 := WithTimeout(time.Second, ApplyContext(s.Wait, ctx))
 			Expect(err3).To(MatchError(context.Canceled))
 		})
 	})
